+++ packages/chatbot-ui/src/components/ui/avatar.tsx
@@ -1,7 +1,7 @@
 import * as AvatarPrimitive from '@radix-ui/react-avatar'
 import type * as React from 'react'
 
-import { cn } from '#/lib/utils'
+import { cn } from '#/components/ui/lib/utils'
 
 function Avatar({ className, ...props }: React.ComponentProps<typeof AvatarPrimitive.Root>) {
   return (
