+++ packages/chatbot-ui/src/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import type * as React from 'react'
 
-import { cn } from '#/lib/utils'
+import { cn } from '#/components/ui/lib/utils'
 
 function Card({ className, ...props }: React.ComponentProps<'div'>) {
   return (
