+++ packages/chatbot-ui/src/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import type * as React from 'react'
 
-import { cn } from '#/lib/utils'
+import { cn } from '#/components/ui/lib/utils'
 
 function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
   return (
