+++ packages/chatbot-ui/src/components/ui/lib/utils.ts
