+++ packages/chatbot-ui/src/components/ui/separator.tsx
@@ -1,7 +1,7 @@
 import * as SeparatorPrimitive from '@radix-ui/react-separator'
 import type * as React from 'react'
 
-import { cn } from '#/lib/utils'
+import { cn } from '#/components/ui/lib/utils'
 
 function Separator({
   className,
